requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
moviepy==1.0.3
scikit-image==0.21.0
Werkzeug==2.3.7